% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.MT.variant.R
\name{DP.MT.variant}
\alias{DP.MT.variant}
\title{Title calculating the distance for each MT variant by Dirichlet process Gaussian mixture}
\usage{
DP.MT.variant(x, iteration = 2000)
}
\arguments{
\item{x}{the VAF vector of one MT variant}
}
\value{
the mean values of all peaks for one MT variants
}
\description{
Title calculating the distance for each MT variant by Dirichlet process Gaussian mixture
}
