% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMut_detect.R
\name{ErrorMut_detect}
\alias{ErrorMut_detect}
\title{Title check the sequencing error for each MT variant, and labeled each MT variant with "detected" or "Error"}
\usage{
ErrorMut_detect(mut, e = 0.001/3, FPR = 5e-07)
}
\arguments{
\item{mut}{a vector with ID,position, s_reads,ave_BQ, t_reads, AF, sample info.}

\item{e}{is sequencing error rate, the default value is 0.001/3}

\item{FPR}{is false-positive rate (FPR), the default value is 0.0000005}
}
\value{
a vector of one MT variant with "detected" or "Error" label
}
\description{
Title check the sequencing error for each MT variant, and labeled each MT variant with "detected" or "Error"
}
