% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMutmatrix_detect.R
\name{ErrorMutmatrix_detect}
\alias{ErrorMutmatrix_detect}
\title{Title find and remove the mutations that likely belong to sequencing error and construct the new matrix}
\usage{
ErrorMutmatrix_detect(MT_matrix, FPR = 5e-07)
}
\arguments{
\item{FPR}{is false-positive rate (FPR), the default value is 0.0000005}

\item{mut}{a data frame, row is the variant, cloum is the ID,position, s_reads,ave_BQ, t_reads, AF, sample info.}

\item{e}{is sequencing error rate, the default value is 0.001/3}
}
\value{
matrix of all MT variants with "detected" or "Error" label
}
\description{
Title find and remove the mutations that likely belong to sequencing error and construct the new matrix
}
