% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT.feature.distance.R
\name{MT.feature.distance}
\alias{MT.feature.distance}
\title{Title calculating distance of all MT variants by Dirichlet process Gaussian mixture}
\usage{
MT.feature.distance(data, iteration = 2000)
}
\arguments{
\item{data}{a VAF matrix for all MT variants, row is VAF, col is Cell}
}
\value{
a list of distance for each MT variant
}
\description{
Title calculating distance of all MT variants by Dirichlet process Gaussian mixture
}
