% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT.feature.selection.R
\name{MT.feature.selection}
\alias{MT.feature.selection}
\title{Title Selecting the informative MT variants}
\usage{
MT.feature.selection(
  distance_list,
  MT_variants_detected_matrix,
  dis_cutoff = 0.02,
  sample_type = 1,
  size = 5
)
}
\arguments{
\item{distance_list}{the distance list of all MT variants generated by MT.feature.distance}

\item{MT_variants_detected_matrix}{the VAF matrix of all detected MT variants generated by ErrorMutmatrix_detect}

\item{dis_cutoff}{the cutoff of distance between top two peaks, the default value of dis_cutoff=0.02}

\item{sample_type}{the default value of sample_type = 1}

\item{size}{remove all adjacent mutations smaller than "size", the default value of size = 5}
}
\value{
The informative MT variants
}
\description{
Title Selecting the informative MT variants
}
