% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTmutMatrix_refined.R
\name{MTmutMatrix_refined}
\alias{MTmutMatrix_refined}
\title{Title transform MT variants table to variants*cell/sample matrix}
\usage{
MTmutMatrix_refined(MT_variants)
}
\arguments{
\item{MT_variants:}{the matrix generated by ErrormutMatrix_detect and removed all "Error" MT variants}
}
\value{
a matrix, row is the variant, cloum is the cell/sample
}
\description{
Title transform MT variants table to variants*cell/sample matrix
}
