% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MT.variant.files.10XscRNASeq.R
\name{read.MT.variant.files.10XscRNASeq}
\alias{read.MT.variant.files.10XscRNASeq}
\title{Title read MT variant files from 10X scRNA-seq data}
\usage{
read.MT.variant.files.10XscRNASeq(path_MTvar)
}
\arguments{
\item{path_MTvar}{is the file path of mtSNP output files}
}
\value{
matrix of all MT variants info
}
\description{
Title read MT variant files from 10X scRNA-seq data
}
