% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MT.variant.files.bulkATACseq.R
\name{read.MT.variant.files.bulkATACseq}
\alias{read.MT.variant.files.bulkATACseq}
\title{Title: read MT variants files of bulk ATACseq data}
\usage{
read.MT.variant.files.bulkATACseq(path_MTvar)
}
\arguments{
\item{path_MTvar}{is the file full path of mtSNP output files}
}
\value{
matrix of all MT variants info
}
\description{
Title: read MT variants files of bulk ATACseq data
}
