% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.MT.variant.files.SmartSeq2.R
\name{read.MT.variant.files.SmartSeq2}
\alias{read.MT.variant.files.SmartSeq2}
\title{Title read MT variant files from Smartseq2 data}
\usage{
read.MT.variant.files.SmartSeq2(path_MTvar)
}
\arguments{
\item{path_MTvar}{is the file path of mtSNP output files}
}
\value{
matrix of all MT variants info
}
\description{
Title read MT variant files from Smartseq2 data
}
