% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_adjacent_MT_variants.R
\name{remove_adjacent_MT_variants}
\alias{remove_adjacent_MT_variants}
\title{Title remove all adjacent MT variants}
\usage{
remove_adjacent_MT_variants(x, size = 5)
}
\arguments{
\item{x}{the character vector of MT variants}

\item{size}{you will remove all adjacent mutations smaller than "size"}
}
\value{
the list of MT variants after removing the adjacent MT variants
}
\description{
Title remove all adjacent MT variants
}
